Add `trigger_build` API, rename `trigger` to `trigger_detail`
The API to execute a trigger is not documented on the same page as the
rest of the trigger management API, but this seems the most appropriate
place for it.

The `trigger` method has been renamed to `trigger_detail` to prevent
confusion as to its behaviour, it retrieves details about a trigger, and
doesn't cause a trigger to be activated.